Refactored API getConversations
