+++ frontend/src/app/data/objects/user.ts
@@ -1,7 +1,6 @@
-// User interface
 export interface User {
     id?: number;  // Unique ID of the user
     accessToken: string;  // Access token for authentication
     email: string;  // Email (used as username)
     name: string;  // Display name
-  }
+}
